requirements.txt
torch
torchvision
opencv-python
pyyaml
numpy
Pillow
tqdm